+++ Code/LYRUIMessageInputToolbar.m
@@ -12,25 +12,26 @@
 #import "LYRUIMessagingUtilities.h"
 #import <CoreLocation/CoreLocation.h>
 
-@interface LYRUIMessageInputToolbar () <UITextViewDelegate, CLLocationManagerDelegate>
+@interface LYRUIMessageInputToolbar () <UITextViewDelegate>
 
-@property (nonatomic) CLLocationManager *locationManager;
-@property (nonatomic) NSLayoutConstraint *textViewHeightConstraint;
 @property (nonatomic) NSArray *messageParts;
 @property (nonatomic, copy) NSAttributedString *attributedStringForMessageParts;
+@property (nonatomic) UITextView *dummyTextView;
+@property (nonatomic) CGFloat textViewMaxHeight;
+@property (nonatomic, getter=isTyping) BOOL typing;
 
 @end
 
 @implementation LYRUIMessageInputToolbar
 
 // Compose View Margin Constants
-static CGFloat const LSComposeviewHorizontalMargin = 6;
-static CGFloat const LSComposeviewVerticalMargin = 6;
+static CGFloat const LYRUIHorizontalMargin = 6;
+static CGFloat const LYRUIVerticalMargin = 6;
 
 // Compose View Button Constants
-static CGFloat const LSLeftAccessoryButtonWidth = 40;
-static CGFloat const LSRightAccessoryButtonWidth = 46;
-static CGFloat const LSButtonHeight = 28;
+static CGFloat const LYRUILeftAccessoryButtonWidth = 40;
+static CGFloat const LYRUIRightAccessoryButtonWidth = 46;
+static CGFloat const LYRUIButtonHeight = 28;
 
 - (id)init
 {
@@ -38,36 +39,25 @@ static CGFloat const LSButtonHeight = 28;
     if (self) {
         self.accessibilityLabel = @"Message Input Toolbar";
         self.translatesAutoresizingMaskIntoConstraints = NO;
-        self.autoresizingMask = UIViewAutoresizingFlexibleHeight;
-        
-        // Setup
+        self.autoresizingMask = UIViewAutoresizingFlexibleWidth;
         self.backgroundColor =  LSLighGrayColor();
         self.canEnableSendButton = YES;
 
-        // Initialize the Camera Button
         self.leftAccessoryButton = [[UIButton alloc] init];
-        self.leftAccessoryButton.translatesAutoresizingMaskIntoConstraints = NO;
         self.leftAccessoryButton.accessibilityLabel = @"Camera Button";
-        self.leftAccessoryButton.layer.cornerRadius = 2;
         [self.leftAccessoryButton setImage:[UIImage imageNamed:@"LayerUIKitResource.bundle/camera"] forState:UIControlStateNormal];
         [self.leftAccessoryButton addTarget:self action:@selector(leftAccessoryButtonTapped) forControlEvents:UIControlEventTouchUpInside];
         [self addSubview:self.leftAccessoryButton];
         
-        // Initialize the Text Input View
         self.textInputView = [[LYRUIMessageComposeTextView alloc] init];
-        self.textInputView.translatesAutoresizingMaskIntoConstraints = NO;
-        self.autoresizingMask = UIViewAutoresizingFlexibleHeight;
         self.textInputView.delegate = self;
         self.textInputView.layer.borderColor = [UIColor lightGrayColor].CGColor;
         self.textInputView.layer.borderWidth = 1;
         self.textInputView.layer.cornerRadius = 4.0f;
         self.textInputView.accessibilityLabel = @"Text Input View";
-        self.textInputView.text = LYRUIPlaceHolderText;
         [self addSubview:self.textInputView];
         
-        // Initialize the Send Button
         self.rightAccessoryButton = [[UIButton alloc] init];
-        self.rightAccessoryButton.translatesAutoresizingMaskIntoConstraints = NO;
         self.rightAccessoryButton.accessibilityLabel = @"Send Button";
         self.rightAccessoryButton.titleLabel.font = [UIFont systemFontOfSize:16];
         [self.rightAccessoryButton setTitle:@"SEND" forState:UIControlStateNormal];
@@ -76,32 +66,99 @@ static CGFloat const LSButtonHeight = 28;
         [self.rightAccessoryButton addTarget:self action:@selector(rightAccessoryButtonTapped) forControlEvents:UIControlEventTouchUpInside];
         [self configureSendButtonEnablement];
         [self addSubview:self.rightAccessoryButton];
-    
-        [self setupLayoutConstraints];
 
-        // Default Max Num Lines is 8
+        // Calling sizeThatFits: or contentSize on the displayed UITextView causes the cursor's position to momentarily appear out of place and prevent scrolling to the selected range. So we use another text view for height calculations.
+        self.dummyTextView = [UITextView new];
+
         self.maxNumberOfLines = 8;
     }
     return self;
 }
 
-- (CGSize)intrinsicContentSize
+- (void)layoutSubviews
 {
-    return CGSizeMake(0, (self.textInputView.intrinsicContentSize.height + LSComposeviewVerticalMargin * 2));
+    [super layoutSubviews];
+
+    // We layout the views manually since using Auto Layout seems to cause issues in this context (i.e. an auto height resizing text view in an input accessory view) especially with iOS 7.1.
+    CGRect frame = self.frame;
+    CGRect leftButtonFrame = self.leftAccessoryButton.frame;
+    CGRect rightButtonFrame = self.rightAccessoryButton.frame;
+    CGRect textViewFrame = self.textInputView.frame;
+
+    leftButtonFrame.size.width = LYRUILeftAccessoryButtonWidth;
+    leftButtonFrame.size.height = LYRUIButtonHeight;
+    leftButtonFrame.origin.x = LYRUIHorizontalMargin;
+
+    rightButtonFrame.size.width = LYRUIRightAccessoryButtonWidth;
+    rightButtonFrame.size.height = LYRUIButtonHeight;
+    rightButtonFrame.origin.x = CGRectGetWidth(frame) - CGRectGetWidth(rightButtonFrame) - LYRUIVerticalMargin;
+
+    textViewFrame.origin.x = CGRectGetMaxX(leftButtonFrame) + LYRUIHorizontalMargin;
+    textViewFrame.origin.y = LYRUIVerticalMargin;
+    textViewFrame.size.width = CGRectGetMinX(rightButtonFrame) - CGRectGetMinX(textViewFrame) - LYRUIHorizontalMargin;
+
+    self.dummyTextView.font = self.textInputView.font;
+    self.dummyTextView.attributedText = self.textInputView.attributedText;
+    CGSize fittedTextViewSize = [self.dummyTextView sizeThatFits:CGSizeMake(CGRectGetWidth(textViewFrame), MAXFLOAT)];
+    textViewFrame.size.height = MIN(fittedTextViewSize.height, self.textViewMaxHeight);
+
+    frame.size.height = CGRectGetHeight(textViewFrame) + LYRUIVerticalMargin * 2;
+    frame.origin.y -= frame.size.height - CGRectGetHeight(self.frame);
+
+    leftButtonFrame.origin.y = CGRectGetHeight(frame) - CGRectGetHeight(leftButtonFrame) - LYRUIVerticalMargin;
+
+    rightButtonFrame.origin.y = CGRectGetHeight(frame) - CGRectGetHeight(rightButtonFrame) - LYRUIVerticalMargin;
+
+    self.leftAccessoryButton.frame = leftButtonFrame;
+    self.rightAccessoryButton.frame = rightButtonFrame;
+    self.textInputView.frame = textViewFrame;
+
+    // Setting one's own frame like this is a no-no but seems to be the lesser of evils when working around the layout issues mentioned above.
+    self.frame = frame;
 }
+
+- (void)paste:(id)sender
+{
+    NSArray *images = [UIPasteboard generalPasteboard].images;
+    if (images.count > 0) {
+        for (UIImage *image in images) {
+            [self insertImage:image];
+        }
+        return;
+    }
+    [super paste:sender];
+}
+
+#pragma mark - Public Methods
 
 - (void)setMaxNumberOfLines:(NSUInteger)maxNumberOfLines
 {
     _maxNumberOfLines = maxNumberOfLines;
-    self.textInputView.maxHeight = self.maxNumberOfLines * self.textInputView.font.lineHeight;
+    self.textViewMaxHeight = self.maxNumberOfLines * self.textInputView.font.lineHeight;
+    [self setNeedsLayout];
 }
-
-#pragma mark Public Content Insertion Methods
 
 - (void)insertImage:(UIImage *)image
 {
-    [self.textInputView insertImage:image];
-    [self adjustFrame];
+    UITextView *textView = self.textInputView;
+
+    NSMutableAttributedString *attributedString = [textView.attributedText mutableCopy];
+    NSAttributedString *lineBreak = [[NSAttributedString alloc] initWithString:@"\n" attributes:@{NSFontAttributeName: self.textInputView.font}];
+    if (attributedString.length > 0 && ![textView.text hasSuffix:@"\n"]) {
+        [attributedString appendAttributedString:lineBreak];
+    }
+
+    LYRUIMediaAttachment *textAttachment = [LYRUIMediaAttachment new];
+    textAttachment.image = image;
+    NSMutableAttributedString *attachmentString = [[NSAttributedString attributedStringWithAttachment:textAttachment] mutableCopy];
+    [attachmentString addAttribute:NSFontAttributeName value:textView.font range:NSMakeRange(0, attachmentString.length)];
+    [attributedString appendAttributedString:attachmentString];
+
+    [attributedString appendAttributedString:lineBreak];
+
+    textView.attributedText = attributedString;
+    self.typing = YES;
+    [self setNeedsLayout];
     [self configureSendButtonEnablement];
 }
 
@@ -127,32 +184,6 @@ static CGFloat const LSButtonHeight = 28;
     }];
 }
 
-#pragma mark Compose View Delegate Methods
-
-- (void)leftAccessoryButtonTapped
-{
-    [self.inputToolBarDelegate messageInputToolbar:self didTapLeftAccessoryButton:self.leftAccessoryButton];
-}
-
-- (void)rightAccessoryButtonTapped
-{
-    if ([self.textInputView.text isEqualToString:LYRUIPlaceHolderText]) return;
-    if ([self.inputToolBarDelegate respondsToSelector:@selector(messageInputToolbarDidEndTyping:)]) {
-        [self.inputToolBarDelegate messageInputToolbarDidEndTyping:self];
-    }
-    if (self.textInputView.text.length > 0) {
-        [self.inputToolBarDelegate messageInputToolbar:self didTapRightAccessoryButton:self.rightAccessoryButton];
-        self.rightAccessoryButton.enabled = NO;
-        [self.textInputView removeAttachements];
-        self.textInputView.text = @"";
-        [self.textInputView layoutSubviews];
-        self.messageParts = nil;
-        self.attributedStringForMessageParts = nil;
-        [self configureSendButtonEnablement];
-    }
-    [self adjustFrame];
-}
-
 - (NSArray *)messageParts
 {
     NSAttributedString *attributedString = self.textInputView.attributedText;
@@ -162,6 +193,82 @@ static CGFloat const LSButtonHeight = 28;
     }
     return _messageParts;
 }
+
+#pragma mark - Accessors
+
+- (void)setTyping:(BOOL)typing
+{
+    if (typing == _typing) return;
+    _typing = typing;
+
+    if (typing && [self.inputToolBarDelegate respondsToSelector:@selector(messageInputToolbarDidBeginTyping:)]) {
+        [self.inputToolBarDelegate messageInputToolbarDidBeginTyping:self];
+    } else if (!typing && [self.inputToolBarDelegate respondsToSelector:@selector(messageInputToolbarDidEndTyping:)]) {
+        [self.inputToolBarDelegate messageInputToolbarDidEndTyping:self];
+    }
+}
+
+#pragma mark - Actions
+
+- (void)leftAccessoryButtonTapped
+{
+    [self.inputToolBarDelegate messageInputToolbar:self didTapLeftAccessoryButton:self.leftAccessoryButton];
+}
+
+- (void)rightAccessoryButtonTapped
+{
+    if (self.textInputView.text.length == 0) return;
+    self.typing = NO;
+    [self.inputToolBarDelegate messageInputToolbar:self didTapRightAccessoryButton:self.rightAccessoryButton];
+    self.rightAccessoryButton.enabled = NO;
+    self.textInputView.text = @"";
+    [self setNeedsLayout];
+    self.messageParts = nil;
+    self.attributedStringForMessageParts = nil;
+    [self configureSendButtonEnablement];
+}
+
+#pragma mark - UITextViewDelegate
+
+- (void)textViewDidChange:(UITextView *)textView
+{
+    self.typing = textView.text.length > 0;
+    [self setNeedsLayout];
+    [self configureSendButtonEnablement];
+
+    // Workaround for iOS 7.1 not scrolling bottom line into view when entering text. Note that in textViewDidChangeSelection: if the selection to the bottom line is due to entering text then the calculation of the bottom content offset won't be accurate since the content size hasn't yet been updated. Content size has been updated by the time this method is called so our calculation will work.
+    NSRange end = NSMakeRange(textView.text.length, 0);
+    if (NSEqualRanges(textView.selectedRange, end)) {
+        CGPoint bottom = CGPointMake(0, textView.contentSize.height - CGRectGetHeight(textView.frame));
+        [textView setContentOffset:bottom animated:NO];
+    }
+}
+
+- (void)textViewDidChangeSelection:(UITextView *)textView
+{
+    // Workaround for iOS 7.1 not scrolling bottom line into view. Note that this only works for a selection change not due to text entry (in other words e.g. when using an external keyboard's bottom arrow key). The workaround in textViewDidChange: handles selection changes due to text entry.
+    NSRange end = NSMakeRange(textView.text.length, 0);
+    if (NSEqualRanges(textView.selectedRange, end)) {
+        CGPoint bottom = CGPointMake(0, textView.contentSize.height - CGRectGetHeight(textView.frame));
+        [textView setContentOffset:bottom animated:NO];
+        return;
+    }
+
+    // Workaround for automatic scrolling not occurring in some cases.
+    [textView scrollRangeToVisible:textView.selectedRange];
+}
+
+- (BOOL)textView:(UITextView *)textView shouldInteractWithURL:(NSURL *)URL inRange:(NSRange)characterRange
+{
+    return YES;
+}
+
+- (BOOL)textView:(UITextView *)textView shouldInteractWithTextAttachment:(NSTextAttachment *)textAttachment inRange:(NSRange)characterRange
+{
+    return YES;
+}
+
+#pragma mark - Helpers
 
 - (NSArray *)messagePartsFromAttributedString:(NSAttributedString *)attributedString
 {
@@ -182,186 +289,7 @@ static CGFloat const LSButtonHeight = 28;
     return messageParts;
 }
 
-#pragma mark TextViewDelegate Methods
-
-- (BOOL)textViewShouldBeginEditing:(UITextView *)textView
-{
-    [self adjustFrame];
-    return YES;
-}
-
-- (BOOL)textViewShouldEndEditing:(UITextView *)textView
-{
-    return YES;
-}
-
-- (void)textViewDidChange:(UITextView *)textView
-{
-    [self adjustFrame];
-    [self configureSendButtonEnablement];
-}
-
-- (BOOL)textView:(UITextView *)textView shouldChangeTextInRange:(NSRange)range replacementText:(NSString *)text
-{
-    if (textView.text.length && (text.length == 0 && range.location == 0 && range.length == textView.text.length)) {
-        // user cleared out the text
-        if ([self.inputToolBarDelegate respondsToSelector:@selector(messageInputToolbarDidEndTyping:)]) {
-            [self.inputToolBarDelegate messageInputToolbarDidEndTyping:self];
-        }
-    } else if (text.length) {
-        if ([self.inputToolBarDelegate respondsToSelector:@selector(messageInputToolbarDidBeginTyping:)]) {
-            [self.inputToolBarDelegate messageInputToolbarDidBeginTyping:self];
-        }
-    }
-    return YES;
-}
-
-- (BOOL)textView:(UITextView *)textView shouldInteractWithURL:(NSURL *)URL inRange:(NSRange)characterRange
-{
-    return YES;
-}
-
-- (BOOL)textView:(UITextView *)textView shouldInteractWithTextAttachment:(NSTextAttachment *)textAttachment inRange:(NSRange)characterRange
-{
-    return YES;
-}
-
-- (void)paste:(id)sender
-{
-    //
-}
-
-- (void)adjustFrame
-{
-    [self invalidateIntrinsicContentSize];
-    
-    // Make sure the text view always scrolls to the bottom
-    CGFloat contentHeight = self.textInputView.contentSize.height;
-    CGFloat height = self.textInputView.frame.size.height;
-    [self.textInputView setContentOffset:CGPointMake(0, contentHeight - height)];
-}
-
-- (void)setupLayoutConstraints
-{
-    //**********Camera Button Constraints**********//
-    // Left Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.leftAccessoryButton
-                                                     attribute:NSLayoutAttributeLeft
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:self
-                                                     attribute:NSLayoutAttributeLeft
-                                                    multiplier:1.0
-                                                      constant:LSComposeviewHorizontalMargin]];
-    // Width
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.leftAccessoryButton
-                                                          attribute:NSLayoutAttributeWidth
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:nil
-                                                          attribute:NSLayoutAttributeNotAnAttribute
-                                                         multiplier:1.0
-                                                           constant:LSLeftAccessoryButtonWidth]];
-
-    // Height
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.leftAccessoryButton
-                                                          attribute:NSLayoutAttributeHeight
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:nil
-                                                          attribute:NSLayoutAttributeNotAnAttribute
-                                                         multiplier:1.0
-                                                           constant:LSButtonHeight]];
-    // Bottom Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.leftAccessoryButton
-                                                          attribute:NSLayoutAttributeBottom
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:self
-                                                          attribute:NSLayoutAttributeBottom
-                                                         multiplier:1.0
-                                                           constant:-LSComposeviewVerticalMargin]];
-    
-    //**********Text Input View Constraints**********//
-    // Left Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.textInputView
-                                                          attribute:NSLayoutAttributeLeft
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:self.leftAccessoryButton
-                                                          attribute:NSLayoutAttributeRight
-                                                         multiplier:1.0
-                                                           constant:LSComposeviewHorizontalMargin]];
-    
-    // Right Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.textInputView
-                                                          attribute:NSLayoutAttributeRight
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:self.rightAccessoryButton
-                                                          attribute:NSLayoutAttributeLeft
-                                                         multiplier:1.0
-                                                           constant:-LSComposeviewHorizontalMargin]];
-    // Bottom Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.textInputView
-                                                          attribute:NSLayoutAttributeBottom
-                                                          relatedBy:NSLayoutRelationEqual
-                                                             toItem:self
-                                                          attribute:NSLayoutAttributeBottom
-                                                         multiplier:1.0
-                                                           constant:-LSComposeviewVerticalMargin]];
-    
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.textInputView
-                                                     attribute:NSLayoutAttributeTop
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:self
-                                                     attribute:NSLayoutAttributeTop
-                                                    multiplier:1.0
-                                                      constant:LSComposeviewVerticalMargin]];
-    
-    //**********Send Button Constraints**********//
-    // Width
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.rightAccessoryButton
-                                                     attribute:NSLayoutAttributeWidth
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:nil
-                                                     attribute:NSLayoutAttributeNotAnAttribute
-                                                    multiplier:1.0
-                                                      constant:LSRightAccessoryButtonWidth]];
-    
-    // Right Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.rightAccessoryButton
-                                                     attribute:NSLayoutAttributeRight
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:self
-                                                     attribute:NSLayoutAttributeRight
-                                                    multiplier:1.0
-                                                      constant:-LSComposeviewVerticalMargin]];
-    // Height
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.rightAccessoryButton
-                                                     attribute:NSLayoutAttributeHeight
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:nil
-                                                     attribute:NSLayoutAttributeNotAnAttribute
-                                                    multiplier:1.0
-                                                      constant:LSButtonHeight]];
-    // Bottom Margin
-    [self addConstraint:[NSLayoutConstraint constraintWithItem:self.rightAccessoryButton
-                                                     attribute:NSLayoutAttributeBottom
-                                                     relatedBy:NSLayoutRelationEqual
-                                                        toItem:self
-                                                     attribute:NSLayoutAttributeBottom
-                                                    multiplier:1.0
-                                                      constant:-LSComposeviewVerticalMargin]];
-}
-
-- (void)layoutSubviews
-{
-    [[self superview] addConstraint:[NSLayoutConstraint constraintWithItem:self
-                                                                 attribute:NSLayoutAttributeWidth
-                                                                 relatedBy:NSLayoutRelationEqual
-                                                                    toItem:self.superview
-                                                                 attribute:NSLayoutAttributeWidth
-                                                                multiplier:1.0
-                                                                  constant:0]];
-    [super layoutSubviews];
-}
-
-#pragma mark Send Button Enablement
+#pragma mark - Send Button Enablement
 
 - (void)setCanEnableSendButton:(BOOL)canEnableSendButton
 {
@@ -378,8 +306,9 @@ static CGFloat const LSButtonHeight = 28;
 - (BOOL)shouldEnableSendButton
 {
     if (!self.canEnableSendButton) return NO;
-    if ([self.textInputView.text isEqualToString:LYRUIPlaceHolderText]) return NO;
-    if (self.textInputView.text.length == 0) return NO;
+    NSString *text = self.textInputView.text;
+    NSString *trimmedText = [text stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
+    if (trimmedText.length == 0) return NO;
     return YES;
 }